+++ CHANGES.md
@@ -6,6 +6,7 @@ Changes
 
 ### Unreleased
 
+* 2017-03-08 - Course settings icon will now be displayed on all sites that display the course header.
 * 2017-03-06 - Changed flat navigation nav-drawer menu code to make all items structured uniformly and improve the layout of the menu.
 * 2017-03-02 - Added Font Awesome
 * 2017-02-04 - Setting to be able to add blocks to the footer and set a one up to three columns layout in the footer